\name{mcmc_convert}
\alias{mcmc_convert}
%- Also NEED an '\alias' for EACH other topic documented here.
\title{
Convert from matrix to mcmc
}
\description{
This function converts a matrix to an object of class mcmc.list.
}
\usage{
mcmc_convert(object)
}
%- maybe also 'usage' for other objects documented here.
\arguments{
  \item{object}{
A matrix}
}

\value{
Object of class mcmc.list.
}
\author{
Georgios Kazantzidis, Malgorzata Roos, Sona Hunanyan 
}
