\name{plot.mce.ed}
\alias{plot.mce.ed}
%- Also NEED an '\alias' for EACH other topic documented here.
\title{
Plot of mce.ed 
}
\description{
Provides plots for the empirical determinacy estimates and their Monte Carlo errors.
}
\usage{
plot.mce.ed(object, ranges = NULL, title = NULL)
}
%- maybe also 'usage' for other objects documented here.
\arguments{
  \item{object}{mce.ed. An object of class mce.ed.}
  \item{ranges}{list. A list of ranges for the limits of the y-axis of the plots.}
  \item{title}{Character. The title of the plot. }
}

\value{
Five plots for TED, EDL, EDS, pEDL, and pEDS estimates and their 95 percent bootstrap confidence intervals. 
}

\author{
Georgios Kazantzidis, Malgorzata Roos, Sona Hunanyan,
}
